move table schema merging into helper
